requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pinecone-client
llama-index
tqdm
